+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 3.1.1
+  - Upgrade es-ruby client to support correct content-type
+
+## 3.1.0
+  - Support for full use of query DSL. Added query_template to use full DSL.
+
 ## 3.0.3
   - Fix couple of bugs related to incorrect variable names